+++ libauc/metrics/__init__.py
@@ -1,2 +1,2 @@
-from .metricsk import *
+from .metrics_k import *
 from .metrics import *
